+++ bloomi/lib/screens/auth/signup/sign_discription.dart
@@ -1,3 +1,4 @@
+import 'package:bloomi/components/custom_button.dart';
 import 'package:bloomi/components/custom_image.dart';
 import 'package:bloomi/components/custom_text.dart';
 import 'package:bloomi/utils/util_constant.dart';
@@ -62,30 +63,6 @@ class _SignUpDescriptionState extends State<SignUpDescription> {
             ),
           ),
         ),
-      ),
-    );
-  }
-}
-
-class CustomButton extends StatelessWidget {
-  const CustomButton({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      alignment: Alignment.center,
-      height: 60,
-      width: 500,
-      decoration: BoxDecoration(
-          color: UtilConstants.primaryColor,
-          borderRadius: BorderRadius.circular(25)),
-      child: const CustomText(
-        "Get Started",
-        fontColor: UtilConstants.whiteColor,
-        fontSize: 20,
-        fontWeight: FontWeight.bold,
       ),
     );
   }
